// packages/components/src/molecules/Trophy/types.ts
export enum TrophyRank {
	SECRET = "SECRET",
	SSS = "SSS",
	SS = "SS",
	S = "S",
	AAA = "AAA",
	AA = "AA",
	A = "A",
	B = "B",
	C = "C",
	UNKNOWN = "?"
}

export interface TrophyTheme {
	BACKGROUND: string;
	TITLE: string;
	ICON_CIRCLE: string;
	TEXT: string;
	LAUREL: string;
	SECRET_RANK_1: string;
	SECRET_RANK_2: string;
	SECRET_RANK_3: string;
	SECRET_RANK_TEXT: string;
	NEXT_RANK_BAR: string;
	S_RANK_BASE: string;
	S_RANK_SHADOW: string;
	S_RANK_TEXT: string;
	A_RANK_BASE: string;
	A_RANK_SHADOW: string;
	A_RANK_TEXT: string;
	B_RANK_BASE: string;
	B_RANK_SHADOW: string;
	B_RANK_TEXT: string;
	DEFAULT_RANK_BASE: string;
	DEFAULT_RANK_SHADOW: string;
	DEFAULT_RANK_TEXT: string;
}